chore(front): separate SpeakerQuotaType on event
