+++ client/src/components/event/types/speaker-quota.ts
@@ -1,0 +1,8 @@
+// 登壇枠の型
+export type SpeakerQuotaType = {
+  id: string;
+  name: string;
+  time: number;
+  currentCount: number;
+  total: number;
+};
